Some design changes

+++ taxi-website-react/src/api/delete-user.js
@@ -2,13 +2,22 @@ import axios from "axios";
 
 const DeleteUser = () => {
   const deleteUser = (userID, fetchData) => {
-    return axios.delete(`http://localhost/taxi-website-project/taxi-website-php/delete_user.php?userID=${userID}`)
-      .then(() => {
-        fetchData();
-      })
-      .catch(error => {
-        console.error('Error deleting user:', error);
-      });
+    return new Promise((resolve, reject) => {
+      const isConfirmed = window.confirm('Are you sure you want to delete this user?');
+      if (isConfirmed) {
+        axios.delete(`http://localhost/taxi-website-project/taxi-website-php/delete_user.php?userID=${userID}`)
+          .then(() => {
+            fetchData();
+            resolve();
+          })
+          .catch(error => {
+            console.error('Error deleting user:', error);
+            reject(error);
+          });
+      } else {
+        resolve();
+      }
+    });
   };
 
   return { deleteUser };
